# Backend/src/docs/SwaggerDocs.yml
openapi: 3.0.0
info: 
  title: Task Dashboard API
  description: |
    API to query users, tasks, and continuously run health checks to ensure container is healthy,
    also adds data into database.
servers:
  - url: http://localhost:3000
    description: Server bind locally into port 3000.
paths:
...
  /task:
    get:
      tags:
        - Task
      summary: Get All Messages
      operationId: getAllTask
      responses:
        "200":
          description: Successful response with task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      tags:
        - Task
      summary: Add a new task
      operationId: postTask
      responses:
        "201":
          description: Task has been added
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
...
components:
  schemas:
...
    Task:
      type: array
      items:
        type: object
        properties:
          taskName:
            type: string
          taskDescription:
            type: string
          taskProgress:
            type: string
            oneOf:
              - NOTSTARTED
              - STARTED
              - COMPLETED
          taskDueDate:
            type: Date